split install to its own test suite
