+++ tests/Install/AdminTest.php
@@ -16,16 +16,30 @@
  * @noinspection PhpUnhandledExceptionInspection
  */
 
-namespace Tests\Feature_v2\Install;
+namespace Tests\Install;
 
 use App\Http\Middleware\AdminUserStatus;
 use App\Http\Middleware\InstallationStatus;
 use Illuminate\Foundation\Testing\DatabaseTransactions;
 use Tests\AbstractTestCase;
+use Tests\Traits\RequiresEmptyUsers;
 
 class AdminTest extends AbstractTestCase
 {
+	use RequiresEmptyUsers;
 	use DatabaseTransactions;
+
+	public function setUp(): void
+	{
+		parent::setUp();
+		$this->setUpRequiresEmptyUsers();
+	}
+
+	public function tearDown(): void
+	{
+		$this->tearDownRequiresEmptyUsers();
+		parent::tearDown();
+	}
 
 	public function testGet(): void
 	{
